Minimize diff.
